backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-docx==1.1.2
python-multipart==0.0.20
beautifulsoup4==4.12.3
requests==2.32.3
openai==1.58.1
reportlab==4.2.5
python-dotenv==1.0.1
pydantic==2.10.3
aiofiles==24.1.0
lxml==5.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
anthropic
google-generativeai
requests
